git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/branches/soc2010/lighter2@34331 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ apps/tools/lighter2/object.cpp
@@ -285,7 +285,7 @@ namespace lighter
 	allPrimitives.SetCapacity (allPrimitives.GetSize() + factPrims.GetSize());
 	for (i = 0; i < factPrims.GetSize(); i++)
 	{
-	  Primitive newPrim (vertexData, (uint)j);
+	  Primitive newPrim (vertexData, j);
 	  
 	  Primitive& prim = allPrimitives[allPrimitives.Push (newPrim)];
 	  //prim.SetOriginalPrimitive (&factPrims[i]);
